requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
idna==2.9
image==1.5.28
Pillow==7.0.0
pytz==2019.3
requests==2.23.0
selenium==3.141.0
sqlparse==0.3.1
urllib3==1.25.8
